Provide high level Assertions classes as entry points
This provides an easy way to use the assertions and will be all that most developers need. This does produce a limitation though. By generating the code, we need all of the `SubjectFactory`s to be exposed in a predictable way, so providing custom anf human-friendly static factory methods on each Subject is not possible.