Replace type() static factories with public constructors (#19)
Truth recommends named factory methods like intents() or views(), but that adds a lot of maintenance work. We had the type() factory because the original AssertJ had something similar, and because in Java 7 this seemed neater than constructor invocations.

With Java 8, the type() factories do not provide better readability in tests than constructor references. Additionally, we provide the Assertions classes that make it so most library consumers never need to reference the Subjects directly. Because of this, we're removing the type() factories.